requirements.txt
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
python-dotenv==1.0.1
streamlit==1.32.2
virtualenv==20.25.1
youtube-transcript-api==0.6.2